Add cluster/workers support to server.js

+++ root/server.js
@@ -7,6 +7,7 @@
 
 require('babel-core/register');
 
+const cluster = require('cluster');
 const fs = require('fs');
 const Raven = require('raven');
 
@@ -20,27 +21,40 @@ const yargs = require('yargs')
     alias: 's',
     default: DBDefs.RENDERER_SOCKET,
     describe: 'UNIX socket path',
+  })
+  .option('workers', {
+    alias: 'w',
+    default: process.env.RENDERER_WORKERS || 1,
+    describe: 'Number of workers to spawn',
   });
 
 Raven.config(DBDefs.SENTRY_DSN).install();
 
 const SOCKET_PATH = yargs.argv.socket;
 
-const socketServer = createServer(SOCKET_PATH);
+if (cluster.isMaster) {
+  const workers = yargs.argv.workers;
+  for (let i = 0; i < workers; i++) {
+    cluster.fork();
+  }
+  console.log(`server.js listening on ${SOCKET_PATH} (pid ${process.pid})`);
+} else {
+  const socketServer = createServer(SOCKET_PATH);
 
-const cleanup = Raven.wrap(function () {
-  socketServer.close(function () {
-    fs.unlinkSync(SOCKET_PATH);
-    process.exit();
+  const cleanup = Raven.wrap(function () {
+    socketServer.close(function () {
+      fs.unlinkSync(SOCKET_PATH);
+      process.exit();
+    });
   });
-});
 
-const hup = Raven.wrap(function () {
-  clearRequireCache();
-  gettext.clearHandles();
-  DBDefs.reload();
-});
+  const hup = Raven.wrap(function () {
+    clearRequireCache();
+    gettext.clearHandles();
+    DBDefs.reload();
+  });
 
-process.on('SIGINT', cleanup);
-process.on('SIGTERM', cleanup);
-process.on('SIGHUP', hup);
+  process.on('SIGINT', cleanup);
+  process.on('SIGTERM', cleanup);
+  process.on('SIGHUP', hup);
+}
